utility script for updating server
  cp bin/config_sample.yml bin/config.yml

Edit bin/config.yml to add your username, the server host,
and the path on the server to deploy to.

Then run: bin/update-server.sh

You will need an ssh account on the server you are
deploying to. In addition you'll need to create the
directory and do the initial git clone on the server.